inkandbridge
Modified the ink color to match the spirit orbs. Changed the liquid script so that it decreases when spirit orbs are used. Changed the script for the bridge so that it utilizes available spirit orbs.
