--- Mo-TechBuild/Assets/Ciconia Studio/Shaders/Glass Shaders/Builtin-Glass Shaders.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 5c14bea9d1f67a24f9acc628b612b798
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
