--- Mo-TechBuild/Assets/Ciconia Studio/Shaders/Glass Shaders/URP - Glass Shaders_v2019 to 2021 LTS.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1991d42b493d43e449d9f8e3f003d8cf
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
