+++ Mo-TechBuild/Assets/Daksh/Inner Jar Fillup/Transparent Material.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 668b6d1f4e7675843a3460548e6c45ec
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
