+++ Mo-TechBuild/Assets/Daksh/transparenthead.mat.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 219be292f65169843a3b1fb70b3b4e8f
+NativeFormatImporter:
+  externalObjects: {}
+  mainObjectFileID: 2100000
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
