--- Mo-TechBuild/Assets/Malbers Animations/Common/Cinemachine/Cinemachine3.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 59d28ba9db90dc940bb2dba7cdc62f27
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
