--- Mo-TechBuild/Assets/Malbers Animations/Common/Scripts/IK/IK Processors/Unity Animation Rig/Processor UnityAnimationRig 1.0.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 14e4674858fa0bb429e9850923506c19
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
