--- Mo-TechBuild/Assets/Malbers Animations/Common/Shaders/Malbers_HDRP_15.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 81b757c49949f6646b475c0d585fe07e
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
