--- Mo-TechBuild/Assets/Malbers Animations/Common/Shaders/Malbers_HDRP_16.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 9ff08244fd9bd494fa2fe5d61039f515
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
