--- Mo-TechBuild/Assets/Malbers Animations/Common/Shaders/Malbers_HDRP_17.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 1405bba2ee7a85f4dbe1264868f3ea47
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
