--- Mo-TechBuild/Assets/Malbers Animations/Common/Shaders/Malbers_Standard.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 7143e432b398cb448ae9ce03221ae94b
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
