--- Mo-TechBuild/Assets/Malbers Animations/Common/Shaders/Malbers_URP_14.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e7d7a15709679454396be573b8825efd
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
