--- Mo-TechBuild/Assets/Malbers Animations/Common/Shaders/Malbers_URP_16.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 3480e9b46b28f614880b80b777cc9acb
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
