--- Mo-TechBuild/Assets/Malbers Animations/Common/Shaders/Malbers_URP_17.unitypackage.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: afd3795bee854724aa8753261b1f66da
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
