--- Mo-TechBuild/Assets/Materials/Inner Jar Fillup.meta
@@ -1,12 +1,0 @@
-fileFormatVersion: 2
-<<<<<<<< HEAD:Mo-TechBuild/Assets/Prefabs/UI Prefab.meta
-guid: 4860c1c42cb8dcc4d8f95229307df0b7
-========
-guid: 093b03505f329d744be8987fd151e970
->>>>>>>> ebe89634799c4948bbd13558b3162c333ada948c:Mo-TechBuild/Assets/Materials/Inner Jar Fillup.meta
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
