--- Mo-TechBuild/Assets/Materials/Inner Jar Fillup/Liquid Material.mat.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-<<<<<<<< HEAD:Mo-TechBuild/Assets/Prefabs/UI Prefab/Pause UI.prefab.meta
-guid: 365dda94aba382149b564986ff4db59a
-PrefabImporter:
-========
-guid: a6f9dcae9dcf91147a91fb7164a1367d
-NativeFormatImporter:
->>>>>>>> ebe89634799c4948bbd13558b3162c333ada948c:Mo-TechBuild/Assets/Materials/Inner Jar Fillup/Liquid Material.mat.meta
-  externalObjects: {}
-  mainObjectFileID: 2100000
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
