--- Mo-TechBuild/Assets/Materials/Inner Jar Fillup/LiquidEffect.shadergraph.meta
@@ -1,10 +1,0 @@
-fileFormatVersion: 2
-guid: efdc68a583e86eb46975a3236a2418d0
-ScriptedImporter:
-  internalIDToNameTable: []
-  externalObjects: {}
-  serializedVersion: 2
-  userData: 
-  assetBundleName: 
-  assetBundleVariant: 
-  script: {fileID: 11500000, guid: 625f186215c104763be7675aa2d941aa, type: 3}
