--- Mo-TechBuild/Assets/Scripts/Liquid.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: f9d250d8efc7b204cac906e66c6092df
