--- Mo-TechBuild/Assets/Scripts/SphereCollection.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 0abadbb25c4b89e4f8457ada6054362b
